refactor: move InfoCard type, page component file export only a component
